made changes to main and fixed gambler interface

+++ src/main/java/com/github/zipcodewilmington/casino/GamblerInterface.java
@@ -2,7 +2,7 @@ package com.github.zipcodewilmington.casino;
 
 public interface GamblerInterface {
 
-    int placeBet();
+    double placeBet();
 
-    int increaseBet();
+    double increaseBet();
 }
